LLM/requirements.txt
streamlit
google-generativeai
python-dotenv
snowflake-connector-python